Week_5_Warehousing_Streaming/sales-warehousing-streaming-pipeline/requirements.txt
pandas
apache-airflow
apache-airflow-providers-google
google-cloud-bigquery
google-auth